+++ CHANGELOG.md
@@ -3,6 +3,18 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Add e-mail confirmation before 2fa setup feature
+- Add config value TwoFactorEmailAuthRequired
+
+### Deprecated
+
+- Deprecate the config field ConfirmMethod in favor of MailRouteMethod. See
+  documentation for these config fields to understand how to use them now.
+
 ## [2.1.0] - 2018-10-28
 
 ### Added